requirements.txt
celery
redis
flask
youtube-dl
yt-dlp
ffmpeg-python
google-cloud-speech
google-cloud-storage
python-dotenv
openai
flask-cors
flask_sqlalchemy
flask_jwt_extended
Werkzeug
google-cloud-secret-manager
psycopg2